Implement NVDataset repartition
NVDataset still does not support file splitting, so repartitioning distributes
whole files across the specified number of partitions. Asking for more
partitions than the number of input files will result in an error.
The repartitioning algorithm tries to spread the files evenly across the
partitions, but since file splitting is not supported skew in the input
can carry over to the NVDataset partitions even after repartitioning.