CrossValidator: trigger GC to free booster
if booster is not disposed, the GPU memory will be leaked because booster holds
native pointer. But the booster finalize is depending on GC, which is not
triggered frequently in GPU situation which is not consuming much JVM memory.

This MR is try to notify JVM to GC after task completes. So when GC happens,
Booster will dispose normally